feat: 现在波形追踪只会在开启测厚时跟随通道1-4的B波门和测厚波门

+++ Mutilple/ToolFunc.cpp
@@ -160,7 +160,7 @@ void GenFBO(GLuint& FBO, GLuint& FBOTEX, GLuint& RBO, GLuint& PBO, GLuint& OUTTE
 
     glGenBuffers(1, &PBO);
     glBindBuffer(GL_PIXEL_PACK_BUFFER, PBO);
-    glBufferData(GL_PIXEL_PACK_BUFFER, width * height * sizeof(DWORD), NULL, GL_DYNAMIC_COPY);
+    glBufferData(GL_PIXEL_PACK_BUFFER, static_cast<unsigned long long>(width) * height * sizeof(DWORD), NULL, GL_DYNAMIC_COPY);
     glBindBuffer(GL_PIXEL_PACK_BUFFER, 0);
 
     glGenTextures(1, &OUTTEX);
